refactor testing into different python files for data_set_manager
